fix: find loaders in a path-separator agnostic fashion
I was using Ubuntu hosted in WSL with a project originally created on
Windows. I was actually running the project from Git for Windows (so, in
Windows). This resulted in some issues for craco-less.

Craco-less attempted to find loaders by first ascertaining the path
separator which should be used. So in Ubuntu in WSL, that would be `/`
since that's Linux. However, the project was installed and being run
from Windows, so the standard path separator is (usually) `\` (though,
PowerShell will work just fine with either `/` or `\`).

But, why should craco-less bother to match _exactly_ the path separator
being used to find a loader? It's unnecessary and brittle. Just so long
as the loader name appears between path separators should be good
enough. Heck, even `\<loader-name>/` shouldn't _really_ be a problem.
And even if it were, I'd be less worried about craco-less than something
else in my build toolchain failing as a result, at which time I would
then correct the issue. So again, matching against an exact path
separator is more complex, brittle, and unnecessary.

This commit replaces the searching logic with a function which uses amap
of loader name to regular expressions so that special-case path
separator logic is not required.